+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.32-SNAPSHOT
 
 **Updates**
+* Updated V2 model implementation.
 * Updated create methods to return the new Entity.
 * Added that Entities can create their own selfLink, both absolute and relative.